pyproject.toml
[project]
name = "medical-ai-application"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.136.1",
    "psycopg2-binary>=2.9.12",
    "psycopg[binary]>=3.3.4",
    "python-dotenv>=1.2.2",
    "sqlmodel>=0.0.38",
]
